feature: reset consumer group offsets
